refactor(llm): separate chatbot handler into modules
