+++ llm/app/services/handlers/__init__.py
@@ -1,0 +1,4 @@
+from . import greeting_handler
+from . import chat_handler as chat_summary_handler # 별칭(alias)을 사용해 이전 코드와 호환성 유지
+from . import project_handler
+from . import org_chart_handler
